requirements.txt
django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
streamlit==1.50.0
scikit-learn==1.7.2
pandas==2.3.2
numpy==2.3.3
matplotlib==3.10.6
seaborn==0.13.2
plotly==6.3.0
reportlab==4.4.4
openpyxl==3.1.5
xlrd==2.0.2
requests==2.32.5
joblib==1.5.2




